% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRedcapData.R
\name{redcap_to_r_data_set}
\alias{redcap_to_r_data_set}
\title{Take your redcap outputs and prepare them for analysis}
\usage{
redcap_to_r_data_set(redcap_data_file, redcap_script_file)
}
\arguments{
\item{redcap_data_file}{The .csv file redcap gave you.}

\item{redcap_script_file}{The .r redcap gave you.}
}
\value{
The data set with better labels, etc.
}
\description{
Take your redcap outputs and prepare them for analysis
}
\examples{
examples forthcoming
}
