% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sem}
\alias{sem}
\title{compute standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{numeric vector, a sample of numbers}
}
\value{
the standard error of the mean.
}
\description{
compute standard error of the mean
}
\section{formula}{

This is the formula for the standard error of the mean.

\deqn{ \frac{1}{2} }

}

\examples{
sem(1:5)
}
